spec: fix/add pre-transaction scripts for geo-rep and cli packages
The cli subpackage never had a %pretrans script, this has been added
now.

The %pretrans script for ge-repliaction was never included in the RPM
package because it was disable by a undefined macro (_can_georeplicate).
This macro is not used/set anywhere else and _without_georeplication
should take care of it anyway.

Note: This is a Red Hat Gluster Storage specific patch. Upstream
      packaging guidelines do not allow these kind of 'features'.

Label: DOWNSTREAM ONLY

Change-Id: I16aab5bba72f1ed178f3bcac47f9d8ef767cfcef
Signed-off-by: Niels de Vos <ndevos@redhat.com>
Signed-off-by: Bala.FA <barumuga@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/50491
Reviewed-on: https://code.engineering.redhat.com/gerrit/60138
Tested-by: Milind Changire <mchangir@redhat.com>