cluster/ec: Improve heal info command to handle obvious cases
Problem:
1 - If a brick is down and we see an index entry in
.glusterfs/indices, we should show it in heal info
output as it most certainly needs heal.

2 - The first problem is also not getting handled after
ec_heal_inspect. Even if in ec_heal_inspect, lookup will
mark need_heal as true, we don't handle it properly in
ec_get_heal_info and continue with locked inspect which
takes lot of time.

Solution:
1 - In first case we need not to do any further invstigation.
As soon as we see that a brick is down, we should say that
this index entry needs heal for sure.

2 - In second case, if we have need_heal as _gf_true after
ec_heal_inspect, we should show it as heal requires.

>Change-Id: Ibe7f9d7602cc0b382ba53bddaf75a2a2c3326aa6
>BUG: 1476668
>Signed-off-by: Ashish Pandey <aspandey@redhat.com>
Upstream Patch: https://review.gluster.org/#/c/17923/

BUG: 1509833
Change-Id: Ibe7f9d7602cc0b382ba53bddaf75a2a2c3326aa6
Signed-off-by: Sunil Kumar Acharya <sheggodu@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/125194
Tested-by: RHGS Build Bot <nigelb@redhat.com>