gfapi: set lkowner in glfd
We need a provision to be able to set lkowner (which is
used to distinguish locks maintained by server) in gfapi.
Since the same lk_owner need to be used to be able to
flush the lock while closing the fd, store the lkowner
in the glfd structure itself.

A new API has been added to be able to set lkowner in glfd.

Upstream reference :
1.) This is backport of below mainline fix -
 https://review.gluster.org/#/c/18429
 https://review.gluster.org/#/c/18522/

2.) 3.12 branch
https://review.gluster.org/#/c/18524/

>Change-Id: I67591d6b9a89c20b9617d52616513ff9e6c06b47
>BUG: 1501956
>Signed-off-by: Soumya Koduri <skoduri@redhat.com>

Change-Id: I67591d6b9a89c20b9617d52616513ff9e6c06b47
BUG: 1500704
Signed-off-by: Jiffin Tony Thottan <jthottan@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/126599
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Atin Mukherjee <amukherj@redhat.com>