quota: fixes issue in quota.conf when setting large number of limits
Problem: It was not possible to configure more than 7712 quota limits.
This was because a stack buffer of size 131072 was used to read from
quota.conf file. In the new format of quota.conf file each gfid entry
takes 17bytes (16byte gfid + 1 byte type). So, the buf_size was not a
multiple of gfid entry size and as per code this was considered as
corruption.

Solution: make buf size a multiple of gfid entry size

> Change-Id: Id036225505a47a4f6fa515a572ee7b0c958f30ed
> BUG: 1510940
> Patch: https://review.gluster.org/#/c/18695/

BUG: 1511766
Change-Id: Id036225505a47a4f6fa515a572ee7b0c958f30ed
Signed-off-by: Sanoj Unnikrishnan <sunnikri@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/124647
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Sunil Kumar Heggodu Gopala Acharya <sheggodu@redhat.com>