ec: Use tiebreaker_inodelk where necessary
When there are big directories or files that need to be healed,
other shds are stuck on getting lock on self-heal domain for these
directories/files. If there is a tie-breaker logic, other shds
can heal some other files/directories while 1 of the shds is healing
the big file/directory.

Before this patch:
96.67  4890.64 us 12.89 us 646115887.30us 340869 INODELK
After this patch:
40.76  42.35 us   15.09 us 6546.50us 438478 INODELK

 >Fixes gluster/glusterfs#354

Upstream-patch: https://review.gluster.org/18820
BUG: 1562744
Change-Id: Ia995b5576b44f770c064090705c78459e543cc64
Signed-off-by: Pranith Kumar K <pkarampu@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/134280
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Ashish Pandey <aspandey@redhat.com>