geo-rep: Cleanup stale unprocessed xsync changelogs
When geo-replication is in hybrid crawl, it crawls
the file system and generates xsync changelogs.
These changelogs are later processed to sync the
data. If the worker goes to Faulty before processing
all the generated xsync changelogs, those changelogs
remain and is not cleaned up. When the worker
comes back, it will re-crawl and re-generate xsync
changelogs. So this patch cleans up the stale
unprocessed xsync changelogs.

Backport of:
 > Patch: https://review.gluster.org/18983
 > Issue: #376

BUG: 1599037
Change-Id: Ib92920c716c7d27e1eeb4bc4ebaf3efb48e0694d
Signed-off-by: Kotresh HR <khiremat@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/144102
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Atin Mukherjee <amukherj@redhat.com>