tests: Use /dev/urandom instead of /dev/random for dd
If there's not enough entropy in the system then reading /dev/random would take
a significant time since it would take a long time for the /dev/random buffers
to get full as is desired in this dd run.
Milind found that this test file takes almost a 1000 seconds or more to pass
instead of just a minute because of this.

>BUG: 1431955
>Change-Id: I9145b17f77f09d0ab71816ae249c69b8fe14c1a5
>Signed-off-by: Pranith Kumar K <pkarampu@redhat.com>
Upstream Patch: https://review.gluster.org/#/c/19160/

BUG: 1509810
Change-Id: I9145b17f77f09d0ab71816ae249c69b8fe14c1a5
Signed-off-by: Sunil Kumar Acharya <sheggodu@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/127405
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Atin Mukherjee <amukherj@redhat.com>