cluster/dht: Handle single dht child in dht_lookup
This patch limits itself to only handling the case
where no file (data or linkto) exists on the subvol.

Additional cases to be handled:
1. A linkto file was found on the only child subvol. This currently
calls dht_lookup_everywhere which eventually deletes it. It can be
deleted directly as it will not be pointing to a valid subvol.
2. Directory lookups - locking might be unnecessary in some cases.

upstream patch: https://review.gluster.org/19581

> Change-Id: I940ba34531f2aaee1d36fd9ca45ecfd46be662a4
> BUG: 1546620
> Signed-off-by: N Balachandran <nbalacha@redhat.com>

Change-Id: I1cc1b9866fa18fe825847585e6a9b8c92898951a
BUG: 1545570
Signed-off-by: N Balachandran <nbalacha@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/130976
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Shyam Ranganathan <srangana@redhat.com>