rpc: make actor search parallel
Problem:
On a service request, the actor is searched using an exclusive mutex
lock which is not really necessary since most of the time the actor
list is going to be searched and not modified.

Solution:
Use a read-write lock instead of a mutex lock. Only modify operations
on a service need to be done under a write-lock which grants exclusive
access to the code.

NOTE:
This patch has been specifically been taken in to use the pthread
rwlock primitives that rpcsvc_ownthread_reconf() is using in commit
8503ed9b94777d47352f19ebfa844e151352b87f
(rpcsvc: scale rpcsvc_request_handler threads)

mainline:
> BUG: 1509644
> Reviewed-on: https://review.gluster.org/18543
> Reviewed-by: Jeff Darcy <jeff@pl.atyp.us>
> Signed-off-by: Milind Changire <mchangir@redhat.com>

Change-Id: I318026c13bb6e0385dc24018a976229182fc7f79
BUG: 1549497
Signed-off-by: Milind Changire <mchangir@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/131909
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Atin Mukherjee <amukherj@redhat.com>